requirements.main.txt
numpy
pillow
shapely